tutorial manager to seperate file
