+++ maps/map_Another_test.json
@@ -1,1 +1,1 @@
-{"mapName": "Another test", "locked": {"isLocked": true, "unlock": 1}, "deletable": true, "saved": true, "width": 16, "height": 9, "difficulty": 3, "completion": {"total": 10, "completed": false, "time": 0}, "audio": "track3", "backgrounds": {"layer 1": [255, 153, 153], "layer 2": [255, 153, 153], "layer 3": [255, 153, 153], "layer 4": [255, 153, 153], "darkMode": true}, "connections": {"layer 2": [[19, 20], [20, 21], [21, 22], [22, 23], [23, 24], [24, 25], [118, 119], [119, 120], [120, 121], [121, 122], [122, 123], [123, 124], [64, 65], [65, 66], [66, 67], [67, 68], [68, 69], [69, 70], [118, 127], [16, 25], [22, 31], [112, 121], [67, 76], [76, 85], [85, 94], [94, 103]], "layer 1": [[29, 30], [30, 31], [31, 32], [32, 33], [33, 42], [42, 51], [51, 60], [60, 69], [69, 78], [78, 87], [87, 96], [96, 105], [105, 114], [110, 111], [111, 112], [112, 113], [113, 114], [29, 38], [38, 47], [47, 56], [56, 65], [65, 74], [74, 83], [83, 92], [92, 101], [101, 110]], "layer 3": [[9, 18], [18, 27], [27, 36], [36, 45], [45, 54], [54, 55], [55, 64], [64, 73], [72, 73], [72, 81], [81, 90], [90, 99], [99, 108], [108, 117], [117, 126], [126, 127], [127, 128], [128, 129], [129, 138], [138, 139], [139, 140], [131, 140], [131, 132], [132, 133], [133, 134], [80, 89], [89, 98], [98, 107], [107, 116], [116, 125], [125, 134], [79, 80], [61, 70], [70, 79], [61, 62], [17, 26], [26, 35], [35, 44], [44, 53], [53, 62], [14, 15], [15, 16], [16, 17], [5, 14], [3, 4], [4, 5], [3, 12], [9, 10], [10, 11], [11, 12]]}, "transport": {"layer 2": [{"location": 20, "type": "bus"}, {"location": 123, "type": "bus"}, {"location": 67, "type": "taxi"}], "layer 1": [{"location": 38, "type": "metro"}, {"location": 96, "type": "metro"}], "layer 3": [{"location": 70, "type": "tram"}, {"location": 64, "type": "tram"}]}, "stops": {"layer 3": [{"location": 64, "type": "tram"}, {"location": 70, "type": "tram"}, {"location": 16, "type": "tram"}, {"location": 127, "type": "tram"}], "layer 1": [{"location": 31, "type": "metro"}, {"location": 112, "type": "metro"}, {"location": 65, "type": "metro"}, {"location": 69, "type": "metro"}], "layer 2": [{"location": 118, "type": "bus"}, {"location": 25, "type": "bus"}, {"location": 31, "type": "bus"}, {"location": 112, "type": "bus"}]}, "destinations": {"layer 3": [{"location": 4, "type": "airport"}, {"location": 139, "type": "airport"}], "layer 2": [{"location": 19, "type": "office"}, {"location": 124, "type": "office"}, {"location": 103, "type": "office"}], "layer 1": [{"location": 29, "type": "house"}, {"location": 114, "type": "house"}, {"location": 110, "type": "house"}, {"location": 33, "type": "house"}]}, "score": 0, "total": 10, "specials": {"layer 2": [{"location": 123, "type": "noWalkNode"}, {"location": 20, "type": "noWalkNode"}, {"location": 94, "type": "noWalkNode"}, {"location": 68, "type": "noWalkNode"}, {"location": 66, "type": "noWalkNode"}, {"location": 24, "type": "noWalkNode"}, {"location": 119, "type": "noWalkNode"}]}}
+{"mapName": "Another test", "locked": {"isLocked": false, "unlock": 1}, "deletable": true, "saved": true, "width": 16, "height": 9, "difficulty": 3, "completion": {"total": 10, "completed": false, "time": 0}, "audio": "track3", "backgrounds": {"layer 1": [255, 153, 153], "layer 2": [255, 153, 153], "layer 3": [255, 153, 153], "layer 4": [255, 153, 153], "darkMode": true}, "connections": {"layer 2": [[19, 20], [20, 21], [21, 22], [22, 23], [23, 24], [24, 25], [118, 119], [119, 120], [120, 121], [121, 122], [122, 123], [123, 124], [64, 65], [65, 66], [66, 67], [67, 68], [68, 69], [69, 70], [118, 127], [16, 25], [22, 31], [112, 121], [67, 76], [76, 85], [85, 94], [94, 103]], "layer 1": [[29, 30], [30, 31], [31, 32], [32, 33], [33, 42], [42, 51], [51, 60], [60, 69], [69, 78], [78, 87], [87, 96], [96, 105], [105, 114], [110, 111], [111, 112], [112, 113], [113, 114], [29, 38], [38, 47], [47, 56], [56, 65], [65, 74], [74, 83], [83, 92], [92, 101], [101, 110]], "layer 3": [[9, 18], [18, 27], [27, 36], [36, 45], [45, 54], [54, 55], [55, 64], [64, 73], [72, 73], [72, 81], [81, 90], [90, 99], [99, 108], [108, 117], [117, 126], [126, 127], [127, 128], [128, 129], [129, 138], [138, 139], [139, 140], [131, 140], [131, 132], [132, 133], [133, 134], [80, 89], [89, 98], [98, 107], [107, 116], [116, 125], [125, 134], [79, 80], [61, 70], [70, 79], [61, 62], [17, 26], [26, 35], [35, 44], [44, 53], [53, 62], [14, 15], [15, 16], [16, 17], [5, 14], [3, 4], [4, 5], [3, 12], [9, 10], [10, 11], [11, 12]]}, "transport": {"layer 2": [{"location": 20, "type": "bus"}, {"location": 123, "type": "bus"}, {"location": 67, "type": "taxi"}], "layer 1": [{"location": 38, "type": "metro"}, {"location": 96, "type": "metro"}], "layer 3": [{"location": 70, "type": "tram"}, {"location": 64, "type": "tram"}]}, "stops": {"layer 3": [{"location": 64, "type": "tram"}, {"location": 70, "type": "tram"}, {"location": 16, "type": "tram"}, {"location": 127, "type": "tram"}], "layer 1": [{"location": 31, "type": "metro"}, {"location": 112, "type": "metro"}, {"location": 65, "type": "metro"}, {"location": 69, "type": "metro"}], "layer 2": [{"location": 118, "type": "bus"}, {"location": 25, "type": "bus"}, {"location": 31, "type": "bus"}, {"location": 112, "type": "bus"}]}, "destinations": {"layer 3": [{"location": 4, "type": "airport"}, {"location": 139, "type": "airport"}], "layer 2": [{"location": 19, "type": "office"}, {"location": 124, "type": "office"}, {"location": 103, "type": "office"}], "layer 1": [{"location": 29, "type": "house"}, {"location": 114, "type": "house"}, {"location": 110, "type": "house"}, {"location": 33, "type": "house"}]}, "score": 0, "total": 10, "specials": {"layer 2": [{"location": 123, "type": "noWalkNode"}, {"location": 20, "type": "noWalkNode"}, {"location": 94, "type": "noWalkNode"}, {"location": 68, "type": "noWalkNode"}, {"location": 66, "type": "noWalkNode"}, {"location": 24, "type": "noWalkNode"}, {"location": 119, "type": "noWalkNode"}]}}
