+++ maps/map_move_test.json
@@ -1,1 +1,1 @@
-{"mapName": "move test", "options": {"lives": false}, "locked": {"isLocked": true, "unlock": 0}, "deletable": true, "saved": true, "width": 18, "height": 10, "difficulty": 1, "total": 1, "completion": {"total": 10, "completed": false, "time": 0}, "backgrounds": {"layer 1": [246, 246, 238], "layer 2": [246, 246, 238], "layer 3": [246, 246, 238], "layer 4": [246, 246, 238]}, "connections": {"layer 2": [[72, 73], [73, 74], [74, 75]], "layer 1": [[42, 52], [52, 62], [62, 72], [72, 82], [82, 92], [92, 102], [102, 112], [112, 122]]}, "transport": {"layer 1": [{"location": 82, "type": "metro"}], "layer 2": [{"location": 73, "type": "taxi"}], "layer 3": []}, "stops": {"layer 1": [{"location": 72, "type": "metro"}], "layer 2": [{"location": 72, "type": "bus"}], "layer 3": []}, "destinations": {"layer 2": [{"location": 75, "type": "airport"}], "layer 1": [{"location": 122, "type": "house"}]}, "score": 0, "specials": {"layer 2": []}}
+{"mapName": "move test", "options": {"lives": false}, "locked": {"isLocked": false, "unlock": 0}, "deletable": true, "saved": true, "width": 18, "height": 10, "difficulty": 1, "total": 1, "completion": {"total": 10, "completed": false, "time": 0}, "backgrounds": {"layer 1": [246, 246, 238], "layer 2": [246, 246, 238], "layer 3": [246, 246, 238], "layer 4": [246, 246, 238]}, "connections": {"layer 2": [[72, 73], [73, 74], [74, 75]], "layer 1": [[42, 52], [52, 62], [62, 72], [72, 82], [82, 92], [92, 102], [102, 112], [112, 122]]}, "transport": {"layer 1": [{"location": 82, "type": "metro"}], "layer 2": [{"location": 73, "type": "taxi"}], "layer 3": []}, "stops": {"layer 1": [{"location": 72, "type": "metro"}], "layer 2": [{"location": 72, "type": "bus"}], "layer 3": []}, "destinations": {"layer 2": [{"location": 75, "type": "airport"}], "layer 1": [{"location": 122, "type": "house"}]}, "score": 0, "specials": {"layer 2": []}}
