+++ maps/map_size4.json
@@ -1,1 +1,1 @@
-{"mapName": "size4", "locked": {"isLocked": true, "unlock": 3}, "deletable": true, "saved": true, "width": 22, "height": 12, "difficulty": 1, "completion": {"total": 10, "completed": false, "time": 0}, "backgrounds": {"layer 1": [246, 246, 238], "layer 2": [246, 246, 238], "layer 3": [246, 246, 238], "layer 4": [246, 246, 238]}, "connections": {"layer 2": [[0, 12], [12, 24], [24, 36], [36, 48], [48, 60], [60, 72], [72, 84], [84, 96], [96, 108], [108, 120], [120, 132], [132, 144], [144, 156], [156, 168], [168, 180], [180, 192], [192, 204], [204, 216], [216, 228], [228, 240], [240, 252], [252, 253], [253, 254], [254, 255], [255, 256], [256, 257], [257, 258], [258, 259], [259, 260], [260, 261], [261, 262], [262, 263], [11, 23], [23, 35], [35, 47], [47, 59], [59, 71], [71, 83], [83, 95], [95, 107], [107, 119], [119, 131], [131, 143], [143, 155], [155, 167], [167, 179], [179, 191], [191, 203], [203, 215], [215, 227], [227, 239], [239, 251], [251, 263], [0, 1], [1, 2], [2, 3], [3, 4], [4, 5], [5, 6], [6, 7], [7, 8], [8, 9], [9, 10], [10, 11]]}, "transport": {}, "stops": {}, "destinations": {"layer 2": [{"location": 0, "type": "airport"}, {"location": 252, "type": "airport"}]}}
+{"mapName": "size4", "locked": {"isLocked": false, "unlock": 3}, "deletable": true, "saved": true, "width": 22, "height": 12, "difficulty": 1, "completion": {"total": 10, "completed": false, "time": 0}, "backgrounds": {"layer 1": [246, 246, 238], "layer 2": [246, 246, 238], "layer 3": [246, 246, 238], "layer 4": [246, 246, 238]}, "connections": {"layer 2": [[0, 12], [12, 24], [24, 36], [36, 48], [48, 60], [60, 72], [72, 84], [84, 96], [96, 108], [108, 120], [120, 132], [132, 144], [144, 156], [156, 168], [168, 180], [180, 192], [192, 204], [204, 216], [216, 228], [228, 240], [240, 252], [252, 253], [253, 254], [254, 255], [255, 256], [256, 257], [257, 258], [258, 259], [259, 260], [260, 261], [261, 262], [262, 263], [11, 23], [23, 35], [35, 47], [47, 59], [59, 71], [71, 83], [83, 95], [95, 107], [107, 119], [119, 131], [131, 143], [143, 155], [155, 167], [167, 179], [179, 191], [191, 203], [203, 215], [215, 227], [227, 239], [239, 251], [251, 263], [0, 1], [1, 2], [2, 3], [3, 4], [4, 5], [5, 6], [6, 7], [7, 8], [8, 9], [9, 10], [10, 11]]}, "transport": {}, "stops": {}, "destinations": {"layer 2": [{"location": 0, "type": "airport"}, {"location": 252, "type": "airport"}]}}
